Extract words types
