+++ src/lib/components/ArticleButton.svelte
@@ -1,6 +1,8 @@
 <script lang="ts">
+	import type { Articles } from '$lib/types/words';
+
 	export let onClickHandler: any;
-	export let choice: string;
+	export let choice: Articles;
 </script>
 
 <button on:click={onClickHandler}>{choice}</button>
@@ -10,5 +12,9 @@
 		all: unset;
 		background-color: aqua;
 		cursor: pointer;
+		width: 5rem;
+		border-radius: 2rem;
+		text-align: center;
+		padding: 1rem;
 	}
 </style>
