+++ src/lib/types/words.ts
@@ -1,0 +1,11 @@
+export enum Articles {
+	DER = 'DER',
+	DIE = 'DIE',
+	DAS = 'DAS'
+}
+
+export type Word = {
+	text: string;
+	article: Articles;
+	choices: Articles[];
+};
